Finish 4.18-4.22

+++ part4/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -9,36 +11,63 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: "5fec41ab943e933c80eb0338"
+    user: user._id
   })
 
   const savedBlog = await blog.save()
 
-  const user = await User.findById("5fec41ab943e933c80eb0338")
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
   response.json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id)
+    return response.status(401).json({ error: 'token missing or invalid' })
+  else if (!blog)
+    return response.status(400).json({ error: 'the blog with specified id cannot be found' })
+
+
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  } else
+    return response.status(401).json({ error: 'current user does not have permission to delete' })
 })
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id)
+    return response.status(401).json({ error: 'token missing or invalid' })
+  else if (!blog)
+    return response.status(400).json({ error: 'the blog with specified id cannot be found' })
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: "5fec41ab943e933c80eb0338"
+    user: user._id
   }
 
   Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
